JECFA Distiller Dataset

This dataset contains information extracted from various JECFA reports and related sources. It includes details about toxicological monographs, specifications, and other relevant data points.

Format

A data frame with 7 rows and 36 variables:

Report

Character. Report identifier, e.g., "TRS 751-JECFA 30/18".

Report_sourcelink

Character. Link to the source of the report, e.g., "https://apps.who.int/iris/".

Tox.Monograph

Character. Identifier for the toxicological monograph, e.g., "FAS 21-JECFA 30/55".

Tox.Monograph_sourcelink

Character. Link to the toxicological monograph, e.g., "http://www.inchem.org/documents/".

Specification

Character. Specification details, e.g., "FAO Combined Compendium of Food Additive Specifications".

Specification_sourcelink

Character. Link to the specification source, e.g., "https://www.fao.org/food/food-safety-quality/".

Synonyms

Character. Synonyms for the chemical, e.g., "Vitamin E; RRR-Alpha-tocopherol".

CAS.number

Character. CAS number for the chemical, if available.

Functional.Class

Character. Functional class of the chemical, e.g., "Food AdditivesANTIOXIDANTNUTRIENT".

Evaluation.year

Character. Year of evaluation, e.g., "1986".

Chemical.Names

Character. Chemical names, e.g., "(2R,4',R,8'R)-2,5,7,8-Tetra".

JECFA.number

Character. JECFA number, if available.

COE.number

Character. COE number, if available.

FEMA.number

Character. FEMA number, if available.

JECFA_name

Character. Name used by JECFA, e.g., "(+)-alpha-TOCOPHEROL".

URL

Character. URL to additional resources or information.

FAS

Character. FAS identifier, e.g., "21".

type

Character. File type, e.g., ".htm", "html".

host

Character. Host of the document, e.g., "www.inchem.org".

Tox_monograph_abbr

Character. Abbreviation for the toxicological monograph, e.g., "FAS 21".

Report_clean

Character. Cleaned report identifier, e.g., "TRS 751-JECFA 30/18".

ref_id

Integer. Reference ID, e.g., 21.

source

Character. Source of the information, e.g., "FAS".

file

Character. File name, e.g., "21.pdf".

dose-response

Logical. Indicates if dose-response information is available.

dose response

Logical. Duplicate column indicating if dose-response information is available.

modelling

Logical. Indicates if modelling information is available.

modeling

Logical. Duplicate column indicating if modeling information is available.

bmd

Logical. Indicates if benchmark dose (BMD) information is available.

bmr

Logical. Indicates if benchmark response (BMR) information is available.

benchmark dose

Logical. Duplicate column indicating if benchmark dose information is available.

benchmark-dose

Logical. Duplicate column indicating if benchmark dose information is available.

Chemical.Names_conc

Character. Concatenated chemical names.

URL_conc

Character. Concatenated URLs.

ref_distiller

Integer. Reference distiller ID, e.g., 1.

CAS.number_1

Character. Additional CAS number information, if available.

Source

Extracted from JECFA reports and related sources.